requirements_versions.txt
torchsde==0.2.5
einops==0.4.1
transformers==4.30.2
safetensors==0.3.1
accelerate==0.21.0
pyyaml==6.0
Pillow==9.2.0
scipy==1.9.3
tqdm==4.64.1
psutil==5.9.5
numpy==1.23.5
pytorch_lightning==1.9.4
omegaconf==2.2.3
gradio==3.47.1
pygit2==1.12.2
aiohttp==3.8.5
playsound2==1.3.0
diffusers==0.21.4
insightface==0.7.3
onnxruntime-gpu==1.16.1
